Some refactoring after code climate review
